Renaming `elementConfiguration` directory to `elementConfigurationSections`.
