+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/VisualizationConfiguration.tsx
@@ -23,7 +23,7 @@ import { Checkbox } from 'components/graylog';
 import Select from 'components/common/Select';
 import usePluginEntities from 'views/logic/usePluginEntities';
 import { defaultCompare } from 'views/logic/DefaultCompare';
-import VisualizationConfigurationOptions from 'views/components/aggregationwizard/elementConfiguration/VisualizationConfigurationOptions';
+import VisualizationConfigurationOptions from 'views/components/aggregationwizard/elementConfigurationSections/VisualizationConfigurationOptions';
 import { WidgetConfigFormValues } from 'views/components/aggregationwizard/WidgetConfigForm';
 import { TIMESTAMP_FIELD } from 'views/Constants';
