+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/VisualizationConfigurationOptions.tsx
@@ -18,8 +18,8 @@ import * as React from 'react';
 import { Field, getIn, useFormikContext } from 'formik';
 import { ConfigurationField } from 'views/types';
 
+import BooleanField from 'views/components/aggregationwizard/elementConfigurationSections/configurationFields/BooleanField';
 import { VisualizationConfigFormValues } from 'views/components/aggregationwizard/WidgetConfigForm';
-import BooleanField from 'views/components/aggregationwizard/elementConfiguration/configurationFields/BooleanField';
 import { HoverForHelp } from 'components/common';
 
 import InputField from './configurationFields/InputField';
