+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/configurationFields/SelectField.tsx
@@ -16,9 +16,9 @@
  */
 import * as React from 'react';
 
+import { FieldComponentProps } from 'views/components/aggregationwizard/elementConfigurationSections/VisualizationConfigurationOptions';
 import { Input } from 'components/bootstrap';
 import Select from 'components/common/Select';
-import { FieldComponentProps } from 'views/components/aggregationwizard/elementConfiguration/VisualizationConfigurationOptions';
 
 const makeOptions = (options: ReadonlyArray<string | [string, any]>) => {
   return options.map((option) => {
